Deleted index.php, adding index.html & jQuery libs
Instead of putting ui work into php, we're going to do the ui using
jQuery and CSS, with content being loaded via smaller php scripts
that are responsible only for 1 particular data source. This will
keep the ui decoupled from the information source, and facilitate
the development of monitoring daemons or an iPhone client.