Generate random scaffolding data in SQL
Instead of hardcoding the scaffolding data in the .php files, we're
now going to generate it here, pump it into the database, and then
begin treating the .php files as real software

Data is generated simply by running make. The Makefile calls
a "generate_scaffolding_data.pl" perl script, which generates
a SQL file called "scaffolding_data.sql". This file is treated
as an artifact and is not under version control